changes for accepting dict in make_parameters() moved all controlaction mutations in mutations/controlaction property mutation new style some changes in application/application.py (just error fixes, but it is obsolete anyways)
