+++ trompace/__init__.py
@@ -40,6 +40,8 @@ def encode_list(thelist, encoder):
     for item in thelist:
         if isinstance(item, StringConstant):
             yield item.value
+        elif isinstance(item, dict):
+            yield "{"+make_parameters(**item)+"}"
         else:
             yield encoder.encode(item)
 
@@ -64,7 +66,10 @@ def make_parameters(**kwargs):
         elif isinstance(v, datetime.datetime):
             value = f"{{formatted: {encoder.encode(v.isoformat())}}}"
         elif isinstance(v, list):
+            #value = "[{}]".format(", ".join(item for item in encode_list(v, encoder)))
             value = "[{}]".format(", ".join(item for item in encode_list(v, encoder)))
+        elif isinstance(v, dict):
+            value = "{"+make_parameters(**v)+"}"
         else:
             value = encoder.encode(v)
         parts.append("{}: {}".format(k, value))
